refactor: 폴더 구조 변경
screens과 widgets 폴더 추가
